a funtion to make a combination of weights from different methods by each row

adding additional column which is a mix of weights columns by each row

cross_cat2cat(
  df,
  cols = colnames(df)[grepl("^wei_.*_c2c$", colnames(df))],
  weis = rep(1/length(cols), length(cols)),
  na.rm = TRUE
)

Arguments

df
data.frame

cols
character vector default all columns follow regex like "wei_.*_c2c"

weis
numeric vector Default vector the same lenght as cols and with equally spaced values summing to 1.

na.rm
logical if NA should be skipped, Dafault TRUE

Value

data.frame with an additional column wei_cross_c2c

Examples

data(occup)
data(trans)

occup_old = occup[occup$year == 2008,]
occup_new = occup[occup$year == 2010,]

# mix of methods
occup_2_mix = cat2cat(
  data = list(old = occup_old, new = occup_new, cat_var = "code", time_var = "year"),
  mappings = list(trans = trans, direction = "backward"),
  ml = list(method = c("knn", "rf", "lda"),
            features = c("age", "sex", "edu", "exp", "parttime", "salary"),
            args = list(k = 10, ntree = 30))
)
# correlation between ml models
cor(occup_2_mix$old[, c("wei_knn_c2c", "wei_rf_c2c", "wei_lda_c2c", "wei_freq_c2c")])

#>              wei_knn_c2c wei_rf_c2c wei_lda_c2c wei_freq_c2c
#> wei_knn_c2c    1.0000000  0.8513018   0.8395457    0.8927482
#> wei_rf_c2c     0.8513018  1.0000000   0.8800756    0.8618755
#> wei_lda_c2c    0.8395457  0.8800756   1.0000000    0.8847568
#> wei_freq_c2c   0.8927482  0.8618755   0.8847568    1.0000000

# cross all methods and subset one highest probability category for each subject
occup_old_mix_highest1occup_2_mix <- prune_cat2cat(cross_cat2cat(occup_2_mix$old),
                                                   column = "wei_cross_c2c", method = "highest1")