A set of prune methods which will be usefull after transition process

user could specify one from four methods to prune replications

prune_cat2cat(
  df,
  index = "index_c2c",
  column = "wei_freq_c2c",
  method = "nonzero",
  percent = 50
)

Arguments

df
data.frame

index
character default wei_freq_c2c

column
character default index_c2c

method
character one of four available methods: default "nonzero", "highest", "highest1", "morethan"

percent
integer from 0 to 99

Value

data.frame

Details

method - specify method to reduce number of replications

"nonzero" remove nonzero probabilities

"highest" leave only highest probabilities for each subject- accepting ties

"highest1" leave only highest probabilities for each subject- not accepting ties so always one is returned

"morethan" leave rows where a probability is hgher than value specify by percent argument

Examples

data(occup)
data(trans)

occup_old <- occup[occup$year == 2008, ]
occup_new <- occup[occup$year == 2010, ]

occup_2 <- cat2cat(
  data = list(old = occup_old, new = occup_new, cat_var = "code", time_var = "year"),
  mappings = list(trans = trans, direction = "backward"),
  ml = list(
    method = "knn",
    features = c("age", "sex", "edu", "exp", "parttime", "salary"),
    args = list(k = 10)
  )
)

prune_cat2cat(occup_2$old, method = "nonzero")

#> # A tibble: 345,820 x 19
#>       id   age sex     edu   exp district sector parttime salary code  multipier
#>    <int> <dbl> <lgl> <int> <dbl>    <int> <chr>     <dbl>  <dbl> <chr>     <dbl>
#>  1   387  36.2 FALSE     1 13.4         8 P        1      58595. 2321       147.
#>  2   474  54.8 TRUE      1  2.95        8 P        0.0567  2246. 2321       158.
#>  3   475  35.6 FALSE     1  4.81        8 P        0.5     9300. 2321       230.
#>  4   476  59.7 FALSE     1 35.3         8 P        1      55211. 2321       246.
#>  5   477  63.9 TRUE      1 42.0         8 P        1      41199. 2321       222.
#>  6   479  30.5 FALSE     1  6.05        3 P        1      23386. 2321       210.
#>  7   480  47.7 FALSE     1 20.9         8 P        1      52991. 2321       137.
#>  8   481  38.2 FALSE     1 13.4         8 P        1      48874. 2321       107.
#>  9   482  28.2 FALSE     1  3.64        3 P        0.09    3529. 2321       222.
#> 10   578  30.5 TRUE      1  5.54        8 P        0.61   12782. 2321       207.
#> # … with 345,810 more rows, and 8 more variables: year <int>, code4 <chr>,
#> #   index_c2c <int>, g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>,
#> #   wei_naive_c2c <dbl>, wei_ml_c2c <dbl>

prune_cat2cat(occup_2$old, method = "highest")

#> # A tibble: 34,920 x 19
#>       id   age sex     edu   exp district sector parttime salary code  multipier
#>    <int> <dbl> <lgl> <int> <dbl>    <int> <chr>     <dbl>  <dbl> <chr>     <dbl>
#>  1     3  31.0 TRUE      1  7.48        8 O         0.333 1.79e4 2471       240.
#>  2     8  48.0 TRUE      1 24.5         8 O         1     2.62e5 1112       251.
#>  3    13  44.3 FALSE     1 19.3         8 O         0.5   2.96e4 2241       210.
#>  4    14  48.8 FALSE     4 29.8         8 O         1     5.06e4 3431       275.
#>  5    18  26.3 FALSE     2  2.25        8 O         1     2.47e4 3431       223.
#>  6    19  40.4 FALSE     1 21.8         8 O         1     2.70e4 3431       258.
#>  7    33  37.9 TRUE      5 18.7         8 P         1     3.58e4 5159       282.
#>  8    33  37.9 TRUE      5 18.7         8 P         1     3.58e4 5159       282.
#>  9    40  30.0 FALSE     1  6.08        8 P         0.125 4.60e3 2443       265.
#> 10    40  30.0 FALSE     1  6.08        8 P         0.125 4.60e3 2443       265.
#> # … with 34,910 more rows, and 8 more variables: year <int>, code4 <chr>,
#> #   index_c2c <int>, g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>,
#> #   wei_naive_c2c <dbl>, wei_ml_c2c <dbl>

prune_cat2cat(occup_2$old, method = "highest1")

#> # A tibble: 34,176 x 19
#>       id   age sex     edu   exp district sector parttime salary code  multipier
#>    <int> <dbl> <lgl> <int> <dbl>    <int> <chr>     <dbl>  <dbl> <chr>     <dbl>
#>  1     3  31.0 TRUE      1  7.48        8 O         0.333 1.79e4 2471       240.
#>  2     8  48.0 TRUE      1 24.5         8 O         1     2.62e5 1112       251.
#>  3    13  44.3 FALSE     1 19.3         8 O         0.5   2.96e4 2241       210.
#>  4    14  48.8 FALSE     4 29.8         8 O         1     5.06e4 3431       275.
#>  5    18  26.3 FALSE     2  2.25        8 O         1     2.47e4 3431       223.
#>  6    19  40.4 FALSE     1 21.8         8 O         1     2.70e4 3431       258.
#>  7    33  37.9 TRUE      5 18.7         8 P         1     3.58e4 5159       282.
#>  8    33  37.9 TRUE      5 18.7         8 P         1     3.58e4 5159       282.
#>  9    40  30.0 FALSE     1  6.08        8 P         0.125 4.60e3 2443       265.
#> 10    40  30.0 FALSE     1  6.08        8 P         0.125 4.60e3 2443       265.
#> # … with 34,166 more rows, and 8 more variables: year <int>, code4 <chr>,
#> #   index_c2c <int>, g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>,
#> #   wei_naive_c2c <dbl>, wei_ml_c2c <dbl>

prune_cat2cat(occup_2$old, method = "morethan", percent = 90)

#> # A tibble: 3,333 x 19
#>       id   age sex     edu   exp district sector parttime salary code  multipier
#>    <int> <dbl> <lgl> <int> <dbl>    <int> <chr>     <dbl>  <dbl> <chr>     <dbl>
#>  1   888  48.3 FALSE     3 26.2         8 Q         1     51176. 2242       143.
#>  2  1005  41.5 FALSE     3 20.0         8 Q         1     33859. 2242       247.
#>  3  1007  48.7 FALSE     3 26.3         8 Q         1     29159. 2242       264.
#>  4  1037  54.8 FALSE     4 33.9         6 Q         1     31867. 2242       255.
#>  5  1045  50.2 FALSE     4 29.2         8 Q         1     46132. 2242       158.
#>  6  4013  26.2 FALSE     2  3.83        8 Q         1     45997. 2242       269.
#>  7  4352  48.0 FALSE     2 10.0         8 Q         0.5   19713. 2242       207.
#>  8  7536  43.4 FALSE     3 21.9         6 Q         0.917 36022. 2242       143.
#>  9  8361  50.5 FALSE     4 29.2         8 Q         1     44870. 2242       234.
#> 10  8789  34.7 FALSE     3 13.2         8 Q         1     35720. 2242       231.
#> # … with 3,323 more rows, and 8 more variables: year <int>, code4 <chr>,
#> #   index_c2c <int>, g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>,
#> #   wei_naive_c2c <dbl>, wei_ml_c2c <dbl>

prune_cat2cat(occup_2$old, column = "wei_ml_c2c", method = "nonzero")

#> # A tibble: 128,121 x 19
#>       id   age sex     edu   exp district sector parttime salary code  multipier
#>    <int> <dbl> <lgl> <int> <dbl>    <int> <chr>     <dbl>  <dbl> <chr>     <dbl>
#>  1  5418  25.8 FALSE     4  2.62        8 G         0.688 1.85e4 5221       283.
#>  2  5418  25.8 FALSE     4  2.62        8 G         0.688 1.85e4 5221       283.
#>  3  5418  25.8 FALSE     4  2.62        8 G         0.688 1.85e4 5221       283.
#>  4 10802  54.1 FALSE     4 33.1        14 O         1     3.74e4 2471       207.
#>  5 10802  54.1 FALSE     4 33.1        14 O         1     3.74e4 2471       207.
#>  6 16463  36.8 FALSE     1 13.4        14 C         1     1.57e5 1231       175.
#>  7 18796  35.2 TRUE      1  3.18        2 P         1     4.26e4 2311       224.
#>  8 18796  35.2 TRUE      1  3.18        2 P         1     4.26e4 2311       224.
#>  9 18796  35.2 TRUE      1  3.18        2 P         1     4.26e4 2311       224.
#> 10 18796  35.2 TRUE      1  3.18        2 P         1     4.26e4 2311       224.
#> # … with 128,111 more rows, and 8 more variables: year <int>, code4 <chr>,
#> #   index_c2c <int>, g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>,
#> #   wei_naive_c2c <dbl>, wei_ml_c2c <dbl>