Validate if the trans table contains all proper mappings

Validate if the trans table contains all proper mappings

Usage

validate_cover_cats(u_cats_target, mapp)

Arguments

u_cats_target

vector of unique target period categories

mapp

transition (mapping) table process with `get_mappings`, the "to_base" direction is taken.