Automatic transforming of a categorical variable according to a new encoding

Automatic transforming of a categorical variable according to a new encoding

cat2cat(
  data = list(old = NULL, new = NULL, cat_var = NULL, id_var = NULL, time_var = NULL,
    multipier_var = NULL, freqs_df = NULL),
  mappings = list(trans = NULL, direction = NULL),
  ml = list(method = NULL, features = NULL, args = NULL)
)

Arguments

data
list with 4, 5 or 6 named fileds `old` `new` `cat_var` `time_var` and optional `id_var`,`multipier_var`,`freq_df`

mappings
list with 2 named fileds `trans` `direction`

ml
list with 3 named fields `method` `features` `args``

Value

list of 2 data.frames

Details

data args

"old" data.frame

"new" data.frame

"id_var" name of the id variable

"cat_var" name of the caterogical variable

"time_var" name of the time varaiable

"multipier_var" name of the multipier variable - number of replication needed to reproduce the population

"freqs_df" data.frame with 2 columns where first one is category name and second counts which will be used to assess the probabilities.

mappings args

"trans" data.frame with 2 columns - transition table

"direction" direction - "backward or "forward"

ml args

"method" character "knn"

"features" vector of features names

"args" k number of k nearest n

Examples

data(occup)
data(trans)

occup_old <- occup[occup$year == 2008, ]
occup_new <- occup[occup$year == 2010, ]

cat2cat(
  data = list(old = occup_old, new = occup_new, cat_var = "code", time_var = "year"),
  mappings = list(trans = trans, direction = "forward")
)

#> $old
#> # A tibble: 34,176 x 18
#>       id   age sex     edu   exp district sector parttime salary code  multipier
#>    <int> <dbl> <lgl> <int> <dbl>    <int> <chr>     <dbl>  <dbl> <chr>     <dbl>
#>  1     1  49.3 FALSE     1 26.5         8 O         1     9.47e4 2471       223.
#>  2     2  34.0 FALSE     1  7.54        8 O         1     6.76e4 2471       279.
#>  3     3  31.0 TRUE      1  7.48        8 O         0.333 1.79e4 2471       240.
#>  4     4  55.3 TRUE      1 24.8         8 O         1     2.62e5 1112       234.
#>  5     5  34.1 TRUE      1 10.1         8 O         1     8.45e4 2471       254.
#>  6     6  35.2 FALSE     1 11.2         8 O         1     9.40e4 2479       262.
#>  7     7  32.4 FALSE     1  8.50        8 O         1     6.53e4 3431       256.
#>  8     8  48.0 TRUE      1 24.5         8 O         1     2.62e5 1112       251.
#>  9     9  51.0 FALSE     4 32.0         8 O         1     4.95e4 3431       259.
#> 10    10  69.6 FALSE     1 44.6         8 O         1     1.75e5 2422       245.
#> # … with 34,166 more rows, and 7 more variables: year <int>, code4 <chr>,
#> #   index_c2c <int>, g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <dbl>,
#> #   wei_naive_c2c <dbl>
#> 
#> $new
#> # A tibble: 36,850 x 18
#>       id   age sex     edu   exp district sector parttime salary code  multipier
#>    <int> <dbl> <lgl> <int> <dbl>    <int> <chr>     <dbl>  <dbl> <chr>     <dbl>
#>  1 34177  52.5 TRUE      1 29.5         8 O          1    9.39e4 1213…      240.
#>  2 34178  56.1 FALSE     6 35.7         8 O          0.67 2.81e4 9112…      265.
#>  3 34179  28.4 FALSE     1  6.36        8 O          1    4.35e4 4226…      232.
#>  4 34180  43.7 FALSE     5 23.0         8 O          1    3.37e4 9112…      225.
#>  5 34181  55.0 TRUE      6 34.6         6 O          1    7.27e4 8322…      230.
#>  6 34182  54.1 FALSE     4 27.9         8 O          1    4.73e4 3343…      215.
#>  7 34183  62.6 FALSE     8 24.6         8 O          1    2.15e4 9112…      221.
#>  8 34184  31.2 FALSE     1  6.37        8 O          1    7.53e4 2611…      202.
#>  9 34185  40.6 TRUE      1 18.9         8 O          1    1.02e5 2611…      255.
#> 10 34186  25.0 TRUE      5  5.82        8 O          1    2.11e4 9629…      226.
#> # … with 36,840 more rows, and 7 more variables: year <int>, code4 <chr>,
#> #   index_c2c <int>, g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>,
#> #   wei_naive_c2c <dbl>
#> 

cat2cat(
  data = list(old = occup_old, new = occup_new, cat_var = "code", time_var = "year"),
  mappings = list(trans = trans, direction = "forward"),
  ml = list(
    method = "knn",
    features = c("age", "sex", "edu", "exp", "parttime", "salary"),
    args = list(k = 10)
  )
)

#> $old
#> # A tibble: 34,176 x 19
#>       id   age sex     edu   exp district sector parttime salary code  multipier
#>    <int> <dbl> <lgl> <int> <dbl>    <int> <chr>     <dbl>  <dbl> <chr>     <dbl>
#>  1     1  49.3 FALSE     1 26.5         8 O         1     9.47e4 2471       223.
#>  2     2  34.0 FALSE     1  7.54        8 O         1     6.76e4 2471       279.
#>  3     3  31.0 TRUE      1  7.48        8 O         0.333 1.79e4 2471       240.
#>  4     4  55.3 TRUE      1 24.8         8 O         1     2.62e5 1112       234.
#>  5     5  34.1 TRUE      1 10.1         8 O         1     8.45e4 2471       254.
#>  6     6  35.2 FALSE     1 11.2         8 O         1     9.40e4 2479       262.
#>  7     7  32.4 FALSE     1  8.50        8 O         1     6.53e4 3431       256.
#>  8     8  48.0 TRUE      1 24.5         8 O         1     2.62e5 1112       251.
#>  9     9  51.0 FALSE     4 32.0         8 O         1     4.95e4 3431       259.
#> 10    10  69.6 FALSE     1 44.6         8 O         1     1.75e5 2422       245.
#> # … with 34,166 more rows, and 8 more variables: year <int>, code4 <chr>,
#> #   index_c2c <int>, g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <dbl>,
#> #   wei_naive_c2c <dbl>, wei_ml_c2c <dbl>
#> 
#> $new
#> # A tibble: 36,850 x 19
#>       id   age sex     edu   exp district sector parttime salary code  multipier
#>    <int> <dbl> <lgl> <int> <dbl>    <int> <chr>     <dbl>  <dbl> <chr>     <dbl>
#>  1 34177  52.5 TRUE      1 29.5         8 O          1    9.39e4 1213…      240.
#>  2 34178  56.1 FALSE     6 35.7         8 O          0.67 2.81e4 9112…      265.
#>  3 34179  28.4 FALSE     1  6.36        8 O          1    4.35e4 4226…      232.
#>  4 34180  43.7 FALSE     5 23.0         8 O          1    3.37e4 9112…      225.
#>  5 34181  55.0 TRUE      6 34.6         6 O          1    7.27e4 8322…      230.
#>  6 34182  54.1 FALSE     4 27.9         8 O          1    4.73e4 3343…      215.
#>  7 34183  62.6 FALSE     8 24.6         8 O          1    2.15e4 9112…      221.
#>  8 34184  31.2 FALSE     1  6.37        8 O          1    7.53e4 2611…      202.
#>  9 34185  40.6 TRUE      1 18.9         8 O          1    1.02e5 2611…      255.
#> 10 34186  25.0 TRUE      5  5.82        8 O          1    2.11e4 9629…      226.
#> # … with 36,840 more rows, and 8 more variables: year <int>, code4 <chr>,
#> #   index_c2c <int>, g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>,
#> #   wei_naive_c2c <dbl>, wei_ml_c2c <dbl>
#>