A set of prune methods which will be useful after transition process

user could specify one from four methods to prune replications

prune_c2c(
  df,
  index = "index_c2c",
  column = "wei_freq_c2c",
  method = "nonzero",
  percent = 50
)

Arguments

df

data.frame

index

character default wei_freq_c2c

column

character default index_c2c

method

character one of four available methods: default "nonzero", "highest", "highest1", "morethan"

percent

integer from 0 to 99

Value

data.frame

Details

method - specify method to reduce number of replications

"nonzero" remove nonzero probabilities

"highest" leave only highest probabilities for each subject- accepting ties

"highest1" leave only highest probabilities for each subject- not accepting ties so always one is returned

"morethan" leave rows where a probability is higher than value specify by percent argument

Examples

if (FALSE) {
data(occup_small)
data(occup)
data(trans)

occup_old <- occup_small[occup_small$year == 2008, ]
occup_new <- occup_small[occup_small$year == 2010, ]

occup_ml <- cat2cat(
  data = list(old = occup_old, new = occup_new, cat_var = "code", time_var = "year"),
  mappings = list(trans = trans, direction = "backward"),
  ml = list(
    data = occup_new,
    cat_var = "code",
    method = "knn",
    features = c("age", "sex", "edu", "exp", "parttime", "salary"),
    args = list(k = 10)
  )
)

prune_c2c(occup_ml$old, method = "nonzero")
prune_c2c(occup_ml$old, method = "highest")
prune_c2c(occup_ml$old, method = "highest1")
prune_c2c(occup_ml$old, method = "morethan", percent = 90)

prune_c2c(occup_ml$old, column = "wei_knn_c2c", method = "nonzero")
}