A set of prune methods which will be usefull after transition process

user could specify one from four methods to prune replications

prune_cat2cat(
  df,
  index = "index_c2c",
  column = "wei_freq_c2c",
  method = "nonzero",
  percent = 50
)

Arguments

df
data.frame

index
character default wei_freq_c2c

column
character default index_c2c

method
character one of four available methods: default "nonzero", "highest", "highest1", "morethan"

percent
integer from 0 to 99

Value

data.frame

Details

method - specify method to reduce number of replications

"nonzero" remove nonzero probabilities

"highest" leave only highest probabilities for each subject- accepting ties

"highest1" leave only highest probabilities for each subject- not accepting ties so always one is returned

"morethan" leave rows where a probability is hgher than value specify by percent argument

Examples

data(occup)
data(trans)

occup_old <- occup[occup$year == 2008, ]
occup_new <- occup[occup$year == 2010, ]

occup_2 <- cat2cat(
  data = list(old = occup_old, new = occup_new, cat_var = "code", time_var = "year"),
  mappings = list(trans = trans, direction = "backward"),
  ml = list(
    method = "knn",
    features = c("age", "sex", "edu", "exp", "parttime", "salary"),
    args = list(k = 10)
  )
)

prune_cat2cat(occup_2$old, method = "nonzero")

#> # A tibble: 345,820 x 18
#>       id   age sex     edu   exp district parttime salary code  multipier  year
#>    <int> <dbl> <lgl> <int> <dbl>    <int>    <dbl>  <dbl> <chr>     <dbl> <int>
#>  1     1  49.3 FALSE     1 26.5         8        1 94731. 2471       223.  2008
#>  2     1  49.3 FALSE     1 26.5         8        1 94731. 2471       223.  2008
#>  3     1  49.3 FALSE     1 26.5         8        1 94731. 2471       223.  2008
#>  4     1  49.3 FALSE     1 26.5         8        1 94731. 2471       223.  2008
#>  5     1  49.3 FALSE     1 26.5         8        1 94731. 2471       223.  2008
#>  6     1  49.3 FALSE     1 26.5         8        1 94731. 2471       223.  2008
#>  7     1  49.3 FALSE     1 26.5         8        1 94731. 2471       223.  2008
#>  8     2  34.0 FALSE     1  7.54        8        1 67619. 2471       279.  2008
#>  9     2  34.0 FALSE     1  7.54        8        1 67619. 2471       279.  2008
#> 10     2  34.0 FALSE     1  7.54        8        1 67619. 2471       279.  2008
#> # … with 345,810 more rows, and 7 more variables: code4 <chr>, index_c2c <int>,
#> #   g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>, wei_naive_c2c <dbl>,
#> #   wei_knn_c2c <dbl>

prune_cat2cat(occup_2$old, method = "highest")

#> # A tibble: 34,920 x 18
#>       id   age sex     edu   exp district parttime salary code  multipier  year
#>    <int> <dbl> <lgl> <int> <dbl>    <int>    <dbl>  <dbl> <chr>     <dbl> <int>
#>  1     1  49.3 FALSE     1 26.5         8    1     9.47e4 2471       223.  2008
#>  2     2  34.0 FALSE     1  7.54        8    1     6.76e4 2471       279.  2008
#>  3     3  31.0 TRUE      1  7.48        8    0.333 1.79e4 2471       240.  2008
#>  4     4  55.3 TRUE      1 24.8         8    1     2.62e5 1112       234.  2008
#>  5     5  34.1 TRUE      1 10.1         8    1     8.45e4 2471       254.  2008
#>  6     6  35.2 FALSE     1 11.2         8    1     9.40e4 2479       262.  2008
#>  7     7  32.4 FALSE     1  8.50        8    1     6.53e4 3431       256.  2008
#>  8     8  48.0 TRUE      1 24.5         8    1     2.62e5 1112       251.  2008
#>  9     9  51.0 FALSE     4 32.0         8    1     4.95e4 3431       259.  2008
#> 10    10  69.6 FALSE     1 44.6         8    1     1.75e5 2422       245.  2008
#> # … with 34,910 more rows, and 7 more variables: code4 <chr>, index_c2c <int>,
#> #   g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>, wei_naive_c2c <dbl>,
#> #   wei_knn_c2c <dbl>

prune_cat2cat(occup_2$old, method = "highest1")

#> # A tibble: 34,176 x 18
#>       id   age sex     edu   exp district parttime salary code  multipier  year
#>    <int> <dbl> <lgl> <int> <dbl>    <int>    <dbl>  <dbl> <chr>     <dbl> <int>
#>  1     1  49.3 FALSE     1 26.5         8    1     9.47e4 2471       223.  2008
#>  2     2  34.0 FALSE     1  7.54        8    1     6.76e4 2471       279.  2008
#>  3     3  31.0 TRUE      1  7.48        8    0.333 1.79e4 2471       240.  2008
#>  4     4  55.3 TRUE      1 24.8         8    1     2.62e5 1112       234.  2008
#>  5     5  34.1 TRUE      1 10.1         8    1     8.45e4 2471       254.  2008
#>  6     6  35.2 FALSE     1 11.2         8    1     9.40e4 2479       262.  2008
#>  7     7  32.4 FALSE     1  8.50        8    1     6.53e4 3431       256.  2008
#>  8     8  48.0 TRUE      1 24.5         8    1     2.62e5 1112       251.  2008
#>  9     9  51.0 FALSE     4 32.0         8    1     4.95e4 3431       259.  2008
#> 10    10  69.6 FALSE     1 44.6         8    1     1.75e5 2422       245.  2008
#> # … with 34,166 more rows, and 7 more variables: code4 <chr>, index_c2c <int>,
#> #   g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>, wei_naive_c2c <dbl>,
#> #   wei_knn_c2c <dbl>

prune_cat2cat(occup_2$old, method = "morethan", percent = 90)

#> # A tibble: 3,333 x 18
#>       id   age sex     edu   exp district parttime salary code  multipier  year
#>    <int> <dbl> <lgl> <int> <dbl>    <int>    <dbl>  <dbl> <chr>     <dbl> <int>
#>  1     4  55.3 TRUE      1 24.8         8    1     2.62e5 1112       234.  2008
#>  2     8  48.0 TRUE      1 24.5         8    1     2.62e5 1112       251.  2008
#>  3    10  69.6 FALSE     1 44.6         8    1     1.75e5 2422       245.  2008
#>  4    21  52.2 FALSE     3 32.9         3    1     5.92e4 3432       269.  2008
#>  5    34  31.7 TRUE      1  6.35        8    1     2.16e4 2212       261.  2008
#>  6    51  54.5 FALSE     1 30.6         6    1     3.46e4 2441       224.  2008
#>  7    69  55.2 TRUE      1 31.2         3    1     8.55e4 2121       203.  2008
#>  8   170  39.6 FALSE     3 20.1         8    0.708 3.31e4 3432       372.  2008
#>  9   179  57.7 FALSE     1 32.4         8    1     1.11e5 2411       182.  2008
#> 10   181  55.8 FALSE     1 31.3         8    1     8.06e4 2411       166.  2008
#> # … with 3,323 more rows, and 7 more variables: code4 <chr>, index_c2c <int>,
#> #   g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>, wei_naive_c2c <dbl>,
#> #   wei_knn_c2c <dbl>

prune_cat2cat(occup_2$old, column = "wei_knn_c2c", method = "nonzero")

#> # A tibble: 127,233 x 18
#>       id   age sex     edu   exp district parttime salary code  multipier  year
#>    <int> <dbl> <lgl> <int> <dbl>    <int>    <dbl>  <dbl> <chr>     <dbl> <int>
#>  1     1  49.3 FALSE     1 26.5         8    1     94731. 2471       223.  2008
#>  2     1  49.3 FALSE     1 26.5         8    1     94731. 2471       223.  2008
#>  3     1  49.3 FALSE     1 26.5         8    1     94731. 2471       223.  2008
#>  4     1  49.3 FALSE     1 26.5         8    1     94731. 2471       223.  2008
#>  5     2  34.0 FALSE     1  7.54        8    1     67619. 2471       279.  2008
#>  6     2  34.0 FALSE     1  7.54        8    1     67619. 2471       279.  2008
#>  7     2  34.0 FALSE     1  7.54        8    1     67619. 2471       279.  2008
#>  8     3  31.0 TRUE      1  7.48        8    0.333 17863. 2471       240.  2008
#>  9     3  31.0 TRUE      1  7.48        8    0.333 17863. 2471       240.  2008
#> 10     3  31.0 TRUE      1  7.48        8    0.333 17863. 2471       240.  2008
#> # … with 127,223 more rows, and 7 more variables: code4 <chr>, index_c2c <int>,
#> #   g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>, wei_naive_c2c <dbl>,
#> #   wei_knn_c2c <dbl>