A set of prune methods which will be useful after transition process

user could specify one from four methods to prune replications

prune_cat2cat(
  df,
  index = "index_c2c",
  column = "wei_freq_c2c",
  method = "nonzero",
  percent = 50
)

Arguments

df
data.frame

index
character default wei_freq_c2c

column
character default index_c2c

method
character one of four available methods: default "nonzero", "highest", "highest1", "morethan"

percent
integer from 0 to 99

Value

data.frame

Details

method - specify method to reduce number of replications

"nonzero" remove nonzero probabilities

"highest" leave only highest probabilities for each subject- accepting ties

"highest1" leave only highest probabilities for each subject- not accepting ties so always one is returned

"morethan" leave rows where a probability is higher than value specify by percent argument

Examples

data(occup_small)
data(occup)
data(trans)

occup_old <- occup_small[occup_small$year == 2008, ]
occup_new <- occup_small[occup_small$year == 2010, ]

occup_2 <- cat2cat(
  data = list(old = occup_old, new = occup_new, cat_var = "code", time_var = "year"),
  mappings = list(trans = trans, direction = "backward"),
  ml = list(
    method = "knn",
    features = c("age", "sex", "edu", "exp", "parttime", "salary"),
    args = list(k = 10)
  )
)

prune_cat2cat(occup_2$old, method = "nonzero")

#> # A tibble: 16,267 x 18
#>       id   age sex     edu   exp district parttime salary code  multiplier  year
#>    <int> <dbl> <lgl> <int> <dbl>    <int>    <dbl>  <dbl> <chr>      <dbl> <int>
#>  1    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#>  2    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#>  3    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#>  4    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#>  5    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#>  6    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#>  7    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#>  8    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#>  9    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#> 10    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#> # … with 16,257 more rows, and 7 more variables: code4 <chr>, index_c2c <int>,
#> #   g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>, wei_naive_c2c <dbl>,
#> #   wei_knn_c2c <dbl>

prune_cat2cat(occup_2$old, method = "highest")

#> # A tibble: 2,965 x 18
#>       id   age sex     edu   exp district parttime salary code  multiplier  year
#>    <int> <dbl> <lgl> <int> <dbl>    <int>    <dbl>  <dbl> <chr>      <dbl> <int>
#>  1    53  67.4 TRUE      1 44.4         6     1    7.05e4 2311       207.   2008
#>  2    95  53.6 TRUE      6 36.1         8     1    3.44e4 3119       153.   2008
#>  3    95  53.6 TRUE      6 36.1         8     1    3.44e4 3119       153.   2008
#>  4   119  63.6 FALSE     1 39.6         8     1    1.44e5 2311        74.9  2008
#>  5   144  56.6 TRUE      1 32.7         6     1    7.25e4 2413       176.   2008
#>  6   158  47.0 FALSE     5 23.9         8     1    4.80e4 1239       193.   2008
#>  7   158  47.0 FALSE     5 23.9         8     1    4.80e4 1239       193.   2008
#>  8   175  70.0 TRUE      1  3.04        8     0.25 1.96e4 2311       376.   2008
#>  9   183  47.5 TRUE      1 21.0         8     1    7.78e4 3441       171.   2008
#> 10   194  29.7 FALSE     1  2.91        8     0.5  3.12e4 2411       126.   2008
#> # … with 2,955 more rows, and 7 more variables: code4 <chr>, index_c2c <int>,
#> #   g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>, wei_naive_c2c <dbl>,
#> #   wei_knn_c2c <dbl>

prune_cat2cat(occup_2$old, method = "highest1")

#> # A tibble: 2,468 x 18
#>       id   age sex     edu   exp district parttime salary code  multiplier  year
#>    <int> <dbl> <lgl> <int> <dbl>    <int>    <dbl>  <dbl> <chr>      <dbl> <int>
#>  1    53  67.4 TRUE      1 44.4         6     1    7.05e4 2311       207.   2008
#>  2    95  53.6 TRUE      6 36.1         8     1    3.44e4 3119       153.   2008
#>  3   119  63.6 FALSE     1 39.6         8     1    1.44e5 2311        74.9  2008
#>  4   144  56.6 TRUE      1 32.7         6     1    7.25e4 2413       176.   2008
#>  5   158  47.0 FALSE     5 23.9         8     1    4.80e4 1239       193.   2008
#>  6   175  70.0 TRUE      1  3.04        8     0.25 1.96e4 2311       376.   2008
#>  7   183  47.5 TRUE      1 21.0         8     1    7.78e4 3441       171.   2008
#>  8   194  29.7 FALSE     1  2.91        8     0.5  3.12e4 2411       126.   2008
#>  9   207  51.8 TRUE      2 31.0         8     1    6.91e4 2139       207.   2008
#> 10   224  38.1 TRUE      2 18.9         8     1    3.21e5 1211       364.   2008
#> # … with 2,458 more rows, and 7 more variables: code4 <chr>, index_c2c <int>,
#> #   g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>, wei_naive_c2c <dbl>,
#> #   wei_knn_c2c <dbl>

prune_cat2cat(occup_2$old, method = "morethan", percent = 90)

#> # A tibble: 333 x 18
#>       id   age sex     edu   exp district parttime salary code  multiplier  year
#>    <int> <dbl> <lgl> <int> <dbl>    <int>    <dbl>  <dbl> <chr>      <dbl> <int>
#>  1   183  47.5 TRUE      1 21.0         8      1   7.78e4 3441       171.   2008
#>  2   194  29.7 FALSE     1  2.91        8      0.5 3.12e4 2411       126.   2008
#>  3   438  53.8 FALSE     8 18.1         6      1   1.92e4 7432        68.9  2008
#>  4   811  53.7 FALSE     4 34.2         6      1   2.66e4 3311       111.   2008
#>  5   966  53.9 FALSE     1 32.2         6      1   4.27e4 3211       199.   2008
#>  6  1165  48.7 FALSE     4 27.6         8      1   3.23e4 4142       290.   2008
#>  7  1177  58.0 TRUE      1 38.2         8      1   2.04e5 1225       109.   2008
#>  8  1212  39.6 TRUE      6 22.8         3      1   2.10e4 8272       287.   2008
#>  9  1230  31.2 TRUE      6  8.52        8      1   2.38e4 7141       274.   2008
#> 10  1426  31.1 FALSE     1  8.55        3      1   4.94e4 1112       258.   2008
#> # … with 323 more rows, and 7 more variables: code4 <chr>, index_c2c <int>,
#> #   g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>, wei_naive_c2c <dbl>,
#> #   wei_knn_c2c <dbl>

prune_cat2cat(occup_2$old, column = "wei_knn_c2c", method = "nonzero")

#> # A tibble: 8,024 x 18
#>       id   age sex     edu   exp district parttime salary code  multiplier  year
#>    <int> <dbl> <lgl> <int> <dbl>    <int>    <dbl>  <dbl> <chr>      <dbl> <int>
#>  1    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#>  2    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#>  3    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#>  4    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#>  5    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#>  6    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#>  7    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#>  8    53  67.4 TRUE      1  44.4        6        1 70506. 2311        207.  2008
#>  9    95  53.6 TRUE      6  36.1        8        1 34377. 3119        153.  2008
#> 10    95  53.6 TRUE      6  36.1        8        1 34377. 3119        153.  2008
#> # … with 8,014 more rows, and 7 more variables: code4 <chr>, index_c2c <int>,
#> #   g_new_c2c <chr>, wei_freq_c2c <dbl>, rep_c2c <int>, wei_naive_c2c <dbl>,
#> #   wei_knn_c2c <dbl>