Adjusted summary for linear regression when based on replicated dataset

adjusting lm object results according to original number of degree of freedom. The standard errors, t statistics and p values have to be adjusted because of replicated rows.

summary_c2c(x, df_old, df_new = x$df.residual)

Arguments

x

lm object

df_old

integer number of d.f in original dataset. For bigger datasets `nrow` should be sufficient.

df_new

integer number of d.f in dataset with replicated rows, Default: x$df.residual

Value

data.frame with additional columns over a regular summary.lm output, like correct and statistics adjusted by it.

Details

The size of the correction is equal to sqrt(df_new / df_old). Where standard errors are multiplied and t statistics divided by it. In most cases the default df_new value should be used.

Examples

data(occup_small)
data(trans)

occup_old <- occup_small[occup_small$year == 2008, ]
occup_new <- occup_small[occup_small$year == 2010, ]

occup_2 <- cat2cat(
  data = list(old = occup_old, new = occup_new, cat_var = "code", time_var = "year"),
  mappings = list(trans = trans, direction = "backward"),
  ml = list(
    method = "knn",
    features = c("age", "sex", "edu", "exp", "parttime", "salary"),
    args = list(k = 10)
  )
)
#> Please wait, statistical models are evaluated.

# Regression
# we have to adjust size of std as we artificialy enlarge degrees of freedom
lms <- lm(I(log(salary)) ~ age + sex + factor(edu) + parttime + exp, occup_2$old,
  weights = multiplier * wei_freq_c2c
)

summary_c2c(lms, df_old = nrow(occup_old))
#>                  Estimate   Std. Error    t value      Pr(>|t|) correct
#> (Intercept)   8.433891931 0.0253371317 332.866878  0.000000e+00 2.56638
#> age           0.002352042 0.0006506223   3.615065  3.011795e-04 2.56638
#> sexTRUE       0.262604727 0.0074608582  35.197657 1.292765e-261 2.56638
#> factor(edu)2 -0.106556564 0.0138403094  -7.699002  1.449903e-14 2.56638
#> factor(edu)3 -0.397839093 0.0172218889 -23.100781 3.359280e-116 2.56638
#> factor(edu)4 -0.397856337 0.0107650217 -36.958248 3.567513e-287 2.56638
#> factor(edu)5 -0.396250768 0.0147747359 -26.819482 4.453055e-155 2.56638
#> factor(edu)6 -0.636084480 0.0104401145 -60.926964  0.000000e+00 2.56638
#> factor(edu)7 -0.583106106 0.2390526325  -2.439237  1.472892e-02 2.56638
#> factor(edu)8 -0.671015647 0.0155771419 -43.076943  0.000000e+00 2.56638
#> parttime      1.948074658 0.0165225845 117.903749  0.000000e+00 2.56638
#> exp           0.009106280 0.0006293183  14.470072  3.677888e-47 2.56638
#>              std.error_c statistic_c     p.value_c
#> (Intercept)  0.065024700 129.7028961  0.000000e+00
#> age          0.001669744   1.4086245  1.590722e-01
#> sexTRUE      0.019147395  13.7149063  2.599874e-41
#> factor(edu)2 0.035519489  -2.9999464  2.727281e-03
#> factor(edu)3 0.044197906  -9.0013109  4.344982e-19
#> factor(edu)4 0.027627133 -14.4009274  3.295024e-45
#> factor(edu)5 0.037917582 -10.4503174  4.825268e-25
#> factor(edu)6 0.026793298 -23.7404326 2.186075e-112
#> factor(edu)7 0.613499820  -0.9504585  3.419724e-01
#> factor(edu)8 0.039976861 -16.7851011  5.848459e-60
#> parttime     0.042403225  45.9416624  0.000000e+00
#> exp          0.001615070   5.6383207  1.912665e-08